procedure changes in DL flow

+++ app-service/src/main/java/in/gov/abdm/abha/enrollment/services/enrol/document/EnrolUsingDrivingLicence.java
@@ -1,6 +1,7 @@
 package in.gov.abdm.abha.enrollment.services.enrol.document;
 
 import in.gov.abdm.abha.enrollment.constants.AbhaConstants;
+import in.gov.abdm.abha.enrollment.constants.PropertyConstants;
 import in.gov.abdm.abha.enrollment.constants.StringConstants;
 import in.gov.abdm.abha.enrollment.enums.AccountAuthMethods;
 import in.gov.abdm.abha.enrollment.enums.AccountStatus;
@@ -20,6 +21,7 @@ import in.gov.abdm.abha.enrollment.model.hidbenefit.RequestHeaders;
 import in.gov.abdm.abha.enrollment.model.lgd.LgdDistrictResponse;
 import in.gov.abdm.abha.enrollment.model.nepix.VerifyDLRequest;
 import in.gov.abdm.abha.enrollment.model.notification.NotificationResponseDto;
+import in.gov.abdm.abha.enrollment.model.procedure.SaveAllDataRequest;
 import in.gov.abdm.abha.enrollment.services.database.account.AccountService;
 import in.gov.abdm.abha.enrollment.services.database.account_auth_methods.AccountAuthMethodService;
 import in.gov.abdm.abha.enrollment.services.database.hidphraddress.HidPhrAddressService;
@@ -39,12 +41,15 @@ import in.gov.abdm.error.ABDMError;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
 
 import static in.gov.abdm.abha.constant.ABHAConstants.PROVISIONAL;
 import static in.gov.abdm.abha.constant.ABHAConstants.VERIFIED;
@@ -109,6 +114,8 @@ public class EnrolUsingDrivingLicence {
 
     @Autowired
     DeDuplicationService deDuplicationService;
+    @Value(PropertyConstants.ENROLLMENT_IS_TRANSACTION_DL)
+    private boolean isDLTransactionManagementEnable;
 
     public Mono<EnrolByDocumentResponseDto> verifyAndCreateAccount(EnrolByDocumentRequestDto enrolByDocumentRequestDto, RequestHeaders requestHeaders) {
         enrolByDocumentRequestDto.setDocumentId(GeneralUtils.removeSpecialChar(enrolByDocumentRequestDto.getDocumentId()));
@@ -217,39 +224,44 @@ public class EnrolUsingDrivingLicence {
                                 }
                                 accountDto.setCreatedDate(LocalDateTime.now());
                                 accountDto.setKycdob(Common.getDob(accountDto.getDayOfBirth(), accountDto.getMonthOfBirth(), accountDto.getYearOfBirth()));
-                                return accountService.createAccountEntity(null,accountDto,requestHeaders).flatMap(accountDtoResponse -> {
-                                    log.info(NEW_ENROLLMENT_ACCOUNT_CREATED_AND_UPDATED_IN_DB);
-                                    HidPhrAddressDto hidPhrAddressDto = hidPhrAddressService.prepareNewHidPhrAddress(accountDto);
-                                    return hidPhrAddressService.createHidPhrAddressEntity(hidPhrAddressDto).flatMap(phrAddressDto -> {
-                                        if (phrAddressDto != null) {
-                                            log.info(DEFAULT_PHR_ADDRESS_UPDATED_IN_HID_PHR_ADDRESS_TABLE);
-                                            return addDocumentsInIdentityDocumentEntity(accountDtoResponse, enrolByDocumentRequestDto).flatMap(idDocumentResponse -> {
-                                                if (idDocumentResponse != null) {
-                                                    log.info(DL_DOCUMENTS_STORED_IN_ADV_DB);
-                                                    return accountAuthMethodService.addAccountAuthMethods(Collections.singletonList(new AccountAuthMethodsDto(accountDtoResponse.getHealthIdNumber(), AccountAuthMethods.MOBILE_OTP.getValue()))).flatMap(res -> {
-                                                        if (!res.isEmpty()) {
-                                                            log.info(ACCOUNT_AUTH_METHODS_ADDED);
-                                                            return transactionService.deleteTransactionEntity(transactionDto.getTxnId().toString()).flatMap(responseEntity -> {
-                                                                if (!responseEntity.getStatusCode().equals(HttpStatus.OK)) {
-                                                                    log.warn(FAILED_TO_DELETE_TRANSACTION + transactionDto.getTxnId().toString());
-                                                                } else {
-                                                                    log.info(TRANSACTION_DELETED);
-                                                                }
-                                                                return sendSuccessNotificationAndPrepareDLResponse(accountDto, transactionDto.getTxnId().toString(), requestHeaders);
-                                                            });
-                                                        } else {
-                                                            throw new AbhaDBGatewayUnavailableException();
-                                                        }
-                                                    });
-                                                } else {
-                                                    throw new DocumentGatewayUnavailableException();
-                                                }
-                                            });
-                                        } else {
-                                            throw new AbhaDBGatewayUnavailableException();
-                                        }
+                                if(isDLTransactionManagementEnable){
+                                    return callProcedureToCreateDLAccount(accountDto,enrolByDocumentRequestDto,transactionDto,requestHeaders);
+                                }else{
+                                    return accountService.createAccountEntity(null,accountDto,requestHeaders).flatMap(accountDtoResponse -> {
+                                        log.info(NEW_ENROLLMENT_ACCOUNT_CREATED_AND_UPDATED_IN_DB);
+                                        HidPhrAddressDto hidPhrAddressDto = hidPhrAddressService.prepareNewHidPhrAddress(accountDto);
+                                        return hidPhrAddressService.createHidPhrAddressEntity(hidPhrAddressDto).flatMap(phrAddressDto -> {
+                                            if (phrAddressDto != null) {
+                                                log.info(DEFAULT_PHR_ADDRESS_UPDATED_IN_HID_PHR_ADDRESS_TABLE);
+                                                return addDocumentsInIdentityDocumentEntity(accountDtoResponse, enrolByDocumentRequestDto).flatMap(idDocumentResponse -> {
+                                                    if (idDocumentResponse != null) {
+                                                        log.info(DL_DOCUMENTS_STORED_IN_ADV_DB);
+                                                        return accountAuthMethodService.addAccountAuthMethods(Collections.singletonList(new AccountAuthMethodsDto(accountDtoResponse.getHealthIdNumber(), AccountAuthMethods.MOBILE_OTP.getValue()))).flatMap(res -> {
+                                                            if (!res.isEmpty()) {
+                                                                log.info(ACCOUNT_AUTH_METHODS_ADDED);
+                                                                return transactionService.deleteTransactionEntity(transactionDto.getTxnId().toString()).flatMap(responseEntity -> {
+                                                                    if (!responseEntity.getStatusCode().equals(HttpStatus.OK)) {
+                                                                        log.warn(FAILED_TO_DELETE_TRANSACTION + transactionDto.getTxnId().toString());
+                                                                    } else {
+                                                                        log.info(TRANSACTION_DELETED);
+                                                                    }
+                                                                    return sendSuccessNotificationAndPrepareDLResponse(accountDto, transactionDto.getTxnId().toString(), requestHeaders);
+                                                                });
+                                                            } else {
+                                                                throw new AbhaDBGatewayUnavailableException();
+                                                            }
+                                                        });
+                                                    } else {
+                                                        throw new DocumentGatewayUnavailableException();
+                                                    }
+                                                });
+                                            } else {
+                                                throw new AbhaDBGatewayUnavailableException();
+                                            }
+                                        });
                                     });
-                                });
+                                }
+
                             })
                 ));
     }
@@ -360,6 +372,32 @@ public class EnrolUsingDrivingLicence {
             return prepareErolByDLResponse(accountDto, txnId, true, true, AbhaConstants.ACCOUNT_CREATED_SUCCESSFULLY, requestHeaders);
         } else {
             throw new NotificationGatewayUnavailableException();
+        }
+    }
+
+    private Mono<EnrolByDocumentResponseDto> callProcedureToCreateDLAccount(AccountDto accountDto, EnrolByDocumentRequestDto enrolByDocumentRequestDto, TransactionDto transactionDto, RequestHeaders requestHeaders) {
+        List<AccountDto> accountList = new ArrayList<>();
+        List<HidPhrAddressDto> hidPhrAddressDtoList = new ArrayList<>();
+        accountList.add(accountDto);
+        HidPhrAddressDto hidPhrAddressDto = hidPhrAddressService.prepareNewHidPhrAddress(accountDto);
+        hidPhrAddressDtoList.add(hidPhrAddressDto);
+        if (!accountDto.getHealthIdNumber().isEmpty()) {
+            List<AccountAuthMethodsDto> accountAuthMethodsDtos = new ArrayList<>();
+            if (accountDto.getMobile() != null) {
+                accountAuthMethodsDtos.add(new AccountAuthMethodsDto(accountDto.getHealthIdNumber(), AccountAuthMethods.MOBILE_OTP.getValue()));
+            }
+           return addDocumentsInIdentityDocumentEntity(accountDto, enrolByDocumentRequestDto)
+                    .flatMap(idDocumentResponse -> {
+                        if (idDocumentResponse != null) {
+                            log.info("going to call procedure to create account via DL");
+                           return accountService.saveAllData(SaveAllDataRequest.builder().accounts(accountList).hidPhrAddress(hidPhrAddressDtoList).accountAuthMethods(accountAuthMethodsDtos).build())
+                                    .flatMap(v-> sendSuccessNotificationAndPrepareDLResponse(accountDto, transactionDto.getTxnId().toString(), requestHeaders));
+                        }else{
+                            throw new DocumentGatewayUnavailableException();
+                        }
+                    });
+        } else {
+            throw new AbhaDBGatewayUnavailableException();
         }
     }
 }
